requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
itsdangerous==2.1.2
loguru==0.7.0
MarkupSafe==2.1.2
pyModbusTCP==0.2.0
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
RPi.GPIO==0.7.1
python-dotenv==1.0.0

